home page code reorganization and small fixes
